+++ test/integration/default/default_test.rb
@@ -16,3 +16,8 @@ end
 describe port(80), :skip do
   it { should_not be_listening }
 end
+
+describe service('chef-client') do
+  it { should be_enabled }
+  it { should be_running }
+end